Generalizing the keygen, reshare, refresh APIs
