+++ src/ecdsa/dkg_ecdsa.rs
@@ -1,100 +1,7 @@
-use frost_secp256k1::keys::SigningShare;
-use frost_secp256k1::*;
-
-use crate::ecdsa::KeygenOutput;
-use crate::generic_dkg::*;
-use crate::protocol::errors::InitializationError;
-use crate::protocol::internal::{make_protocol, Comms};
-use crate::protocol::{Participant, Protocol};
-use futures::FutureExt;
-
-type E = Secp256K1Sha256;
-
-/// Performs the Ed25519 DKG protocol
-pub fn keygen(
-    participants: &[Participant],
-    me: Participant,
-    threshold: usize,
-) -> Result<impl Protocol<Output = KeygenOutput>, InitializationError> {
-    let ctx = Comms::new();
-    let participants = assert_keygen_invariants(participants, me, threshold)?;
-    let fut =
-        do_keygen(ctx.shared_channel(), participants, me, threshold).map(|x| x.map(Into::into));
-    Ok(make_protocol(ctx, fut))
-}
-
-/// Performs the Ed25519 Reshare protocol
-pub fn reshare(
-    old_participants: &[Participant],
-    old_threshold: usize,
-    old_signing_key: Option<SigningShare>,
-    old_public_key: VerifyingKey,
-    new_participants: &[Participant],
-    new_threshold: usize,
-    me: Participant,
-) -> Result<impl Protocol<Output = KeygenOutput>, InitializationError> {
-    let ctx = Comms::new();
-    let threshold = new_threshold;
-    let (participants, old_participants) = reshare_assertions::<E>(
-        new_participants,
-        me,
-        threshold,
-        old_signing_key,
-        old_threshold,
-        old_participants,
-    )?;
-    let fut = do_reshare(
-        ctx.shared_channel(),
-        participants,
-        me,
-        threshold,
-        old_signing_key,
-        old_public_key,
-        old_participants,
-    )
-    .map(|x| x.map(Into::into));
-    Ok(make_protocol(ctx, fut))
-}
-
-/// Performs the Ed25519 Refresh protocol
-pub fn refresh(
-    old_signing_key: Option<SigningShare>,
-    old_public_key: VerifyingKey,
-    new_participants: &[Participant],
-    new_threshold: usize,
-    me: Participant,
-) -> Result<impl Protocol<Output = KeygenOutput>, InitializationError> {
-    if old_signing_key.is_none() {
-        return Err(InitializationError::BadParameters(format!(
-            "The participant {me:?} is running refresh without an old share",
-        )));
-    }
-    let ctx = Comms::new();
-    let threshold = new_threshold;
-    let (participants, old_participants) = reshare_assertions::<E>(
-        new_participants,
-        me,
-        threshold,
-        old_signing_key,
-        threshold,
-        new_participants,
-    )?;
-    let fut = do_reshare(
-        ctx.shared_channel(),
-        participants,
-        me,
-        threshold,
-        old_signing_key,
-        old_public_key,
-        old_participants,
-    )
-    .map(|x| x.map(Into::into));
-    Ok(make_protocol(ctx, fut))
-}
-
 #[cfg(test)]
 mod test {
-    use super::*;
+    use frost_secp256k1::{Secp256K1Sha256, Secp256K1Group, Group};
+    type E = Secp256K1Sha256;
 
     use crate::ecdsa::test::{assert_public_key_invariant, run_keygen, run_refresh, run_reshare};
     use crate::participants::ParticipantList;
