+++ src/ecdsa/mod.rs
@@ -25,15 +25,6 @@ pub type CoefficientCommitment = frost_core::keys::CoefficientCommitment<Secp256
 pub type Polynomial = crate::crypto::polynomials::Polynomial<Secp256K1Sha256>;
 pub type PolynomialCommitment = crate::crypto::polynomials::PolynomialCommitment<Secp256K1Sha256>;
 
-impl From<crate::generic_dkg::KeygenOutput<Secp256K1Sha256>> for KeygenOutput {
-    fn from(value: crate::generic_dkg::KeygenOutput<Secp256K1Sha256>) -> Self {
-        Self {
-            private_share: value.private_share,
-            public_key: value.public_key,
-        }
-    }
-}
-
 impl ScalarSerializationFormat for Secp256K1Sha256 {
     fn bytes_order() -> BytesOrder {
         BytesOrder::BigEndian
