+++ src/ecdsa/test.rs
@@ -4,11 +4,10 @@ use std::error::Error;
 
 use crate::crypto::hash::test::scalar_hash;
 use crate::ecdsa::{
-    dkg_ecdsa::{keygen, refresh, reshare},
     FullSignature, KeygenOutput,
 };
-use crate::protocol::errors::InitializationError;
-use crate::protocol::{run_protocol, Participant, Protocol};
+use crate::protocol::{run_protocol, errors::InitializationError, Participant, Protocol};
+use crate::{keygen, refresh, reshare};
 
 /// runs distributed keygen
 pub(crate) fn run_keygen(
